fix(schema): H-1 amount type normalization + H-2 missing FKs
Two related schema-tightening migrations, both with strict
idempotency contracts so a re-run against a partially-applied
DB is safe:

H-1 (047_normalize_amount_types.go):
  - deposits.amount:                VARCHAR(65) → NUMERIC(32, 8)
  - coin_chains.min_deposit_amount:  VARCHAR(64) → NUMERIC(32, 8)
  - Why: the application parses both via decimal.NewFromString on
    read; a numeric column gives server-side precision and
    correctness (`deposits.amount > 0` is currently a lexicographic
    compare that returns wrong results for any value < 1)
  - Pre-check: counts rows where amount doesn't match the regex
    `^-?[0-9]+(\.[0-9]+)?$` and RAISE EXCEPTION on N > 0; this
    prevents a silent USING col::numeric that would mask bad data
  - ALTER COLUMN is itself idempotent (PG short-circuits on
    already-NUMERIC column)
  - Down() refuses: NUMERIC→VARCHAR is destructive, no rollback

H-2 (048_add_missing_fks.go):
  - withdrawal_verification.withdrawal_order_id → withdrawal_order(id)
    with ON DELETE NO ACTION (a small bug fix: the table was
    previously orphan-allowed; deleting a withdrawal_order with
    verifications now fails loudly)
  - withdrawal_freeze_log.order_id → withdrawal_order(id)
    with ON DELETE SET NULL (audit logs survive order deletion)
  - address_pool.assigned_user_id → users(id)
    with ON DELETE SET NULL (the address returns to the available
    pool)
  - Pre-check: counts orphans for each FK and RAISE EXCEPTION on
    N > 0; user must manually clean before retry
  - ADD CONSTRAINT wrapped in pg_constraint precheck for idempotency
  - Down() drops the FK constraints (REVERSIBLE, unlike 047)

Defensive application-side change:
  - internal/wallet/deposit/repository.go (UpsertDeposit INSERT)
    and internal/repository/postgres/deposit.go (Create() legacy
    insert) now use $3::numeric on the amount parameter. The
    implicit text→numeric cast in PG assignment context would
    work, but the explicit cast documents the type contract and
    protects against future PG versions tightening implicit-cast
    rules. The Go scan-side requires no change: both DepositRow.Amount
    and CoinChain.MinDepositAmount are typed as string, and PG's
    NUMERIC→text output is a clean string conversion.

Operational handbook:
  - docs/security/H1-H2-NOTES.md: 7-section runbook covering
    precision choice (32,8 vs 65,30), ON DELETE rationale, pre-check
    failure recovery SQL, and open follow-ups

Note: these migrations are STAGED. The Go file
internal/migration/migrations/migrations_test.go (which has the
interface tests for 047/048) was committed in the C-2 commit
because that's where TestMigrationOrder lives. The test file
references types that are only defined in this commit's migration
files; this is fine because git doesn't enforce per-commit
compilability, but reviewers should know to read C-2 and H-1/H-2
together when evaluating the test coverage.

+++ internal/wallet/deposit/repository.go
@@ -160,7 +160,7 @@ func (r *DBRepository) UpsertDeposit(ctx context.Context, tx Tx, d *DepositRow) 
 		    block_height, block_hash, status,
 		    from_address, to_address)
 		 VALUES
-		   ($1, $2, $3, $4, $5,
+		   ($1, $2, $3::numeric, $4, $5,
 		    $6, $7, $8, $9,
 		    $10, $11, $12,
 		    $13, $14, $15,
